Adjusted installation instructions and packages
Purpose

The original installation instruction pushes for the use of yarn. Yarn
has proven to create problems when running babel-register against
certain libraries, as well as surfacing some dependency issues that may
not necessarily be a problem. As a temporary fix we are switching to npm
installs against embedded npm roots.

Changes

1. Added jquery as a peer dependency for a failing bootstrap
installation
2. Adjusted instructions to ask user to use npm rather than yarn